clear_state_hashes as one batch

+++ crates/sui-core/src/authority/authority_per_epoch_store.rs
@@ -1448,20 +1448,6 @@ impl AuthorityPerEpochStore {
             }
         }
 
-        if !keys_to_remove.is_empty() {
-            let mut batch = tables.running_root_state_hash.batch();
-            batch
-                .delete_batch(&tables.running_root_state_hash, keys_to_remove.clone())
-                .expect("db error");
-            batch.write().expect("db error");
-            for key in keys_to_remove {
-                info!(
-                    "Cleared running root state hash for checkpoint {} (after last committed checkpoint {})",
-                    key, last_committed_checkpoint
-                );
-            }
-        }
-
         let mut checkpoint_keys_to_remove = Vec::new();
         for kv in tables
             .state_hash_by_checkpoint
@@ -1473,15 +1459,28 @@ impl AuthorityPerEpochStore {
             }
         }
 
-        if !checkpoint_keys_to_remove.is_empty() {
-            let mut batch = tables.state_hash_by_checkpoint.batch();
-            batch
-                .delete_batch(
-                    &tables.state_hash_by_checkpoint,
-                    checkpoint_keys_to_remove.clone(),
-                )
-                .expect("db error");
+        if !keys_to_remove.is_empty() || !checkpoint_keys_to_remove.is_empty() {
+            let mut batch = self.db_batch()?;
+            if !keys_to_remove.is_empty() {
+                batch
+                    .delete_batch(&tables.running_root_state_hash, keys_to_remove.clone())
+                    .expect("db error");
+            }
+            if !checkpoint_keys_to_remove.is_empty() {
+                batch
+                    .delete_batch(
+                        &tables.state_hash_by_checkpoint,
+                        checkpoint_keys_to_remove.clone(),
+                    )
+                    .expect("db error");
+            }
             batch.write().expect("db error");
+            for key in keys_to_remove {
+                info!(
+                    "Cleared running root state hash for checkpoint {} (after last committed checkpoint {})",
+                    key, last_committed_checkpoint
+                );
+            }
             for key in checkpoint_keys_to_remove {
                 info!(
                     "Cleared checkpoint state hash for checkpoint {} (after last committed checkpoint {})",
